fix: move angleBetween from math.js to coordinate.js to avoid type import cycles
